Create README.md
This app has been developed with java and it is complementary to "My Shop"  that has been developed for seller. With this app , one can order  product  and save  product to buy that product by going to the shop. Buyers can order product with real time. PUSHER has been used as real time communication server.